test: 💍 Rename `DatabaseEloquentHasManyThroughTest` with fixes
